docker/sandbox/tools/t1/black_forest_labs/requirements.txt
dify_plugin>=0.2.0,<0.4.0
pillow
